refactor: Reorganize project folder structure
