+++ src/client/assets/fonts/weathericons-regular-webfont.eot
