+++ src/client/js/services/applicationState.js
