+++ src/client/js/viewUpdaters/dateViewUpdater.js
