+++ src/client/js/viewUpdaters/pictureViewUpdater.js
