+++ src/client/js/viewUpdaters/weatherViewUpdater.js
