+++ src/client/styles/weather-icons/weather-icons-core.scss
@@ -1,5 +1,5 @@
 // Base Variables
-$wi-path: '../../assets/fonts' !default;
+$wi-path: '../assets/fonts' !default;
 $wi-version: '2.0' !default;
 $wi-css-prefix: wi !default;
 $wi-font-family: weathericons !default;
